feat: debug prints

+++ include/vga/print.h
@@ -4,6 +4,6 @@
 #include<stdint.h>
 #include<vga/vga.h>
 
-uint64_t vga_print(vga_buffer_t *vga, const char * buf, uint8_t size);
+uint64_t vga_print(vga_buffer_t *vga, const char * buf, uint64_t size);
 
 #endif
